requirements.txt
Flask
flask-cors
python-dotenv
#libsql-experimental
libsql
requests
